+++ app/listener.go
@@ -7,16 +7,15 @@ import (
 	"regexp"
 	"strconv"
 	"sync"
-	"time"
 
+	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/golang/protobuf/proto"
 	log "github.com/sirupsen/logrus"
-	"github.com/ethereum/go-ethereum/common/hexutil"	
-	"github.com/spf13/viper"
 	"github.com/spf13/pflag"
-	"gitlab.com/ConsenSys/client/fr/core-stack/core.git"
-	tracepb "gitlab.com/ConsenSys/client/fr/core-stack/core.git/protobuf/trace"
-	"gitlab.com/ConsenSys/client/fr/core-stack/infra/ethereum.git/tx-listener"
+	"github.com/spf13/viper"
+	listener "gitlab.com/ConsenSys/client/fr/core-stack/infra/ethereum.git/tx-listener"
+	coreworker "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
+	trace "gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/protos/trace"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-listener.git/app/infra"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-listener.git/app/worker"
 )
@@ -29,7 +28,7 @@ type ListenerHandler struct {
 	defaultPosition *StartingPosition
 
 	cleanOnce *sync.Once
-	worker    *core.Worker
+	worker    *coreworker.Worker
 	logger    *log.Entry
 
 	cfg listener.Config
@@ -41,11 +40,14 @@ type ListenerHandler struct {
 func (l *ListenerHandler) Setup() {
 	l.worker = worker.CreateWorker(l.app.infra)
 
+	// Add partitioning so chains can be treated in parallel
+	l.worker.Partitionner(func(msg interface{}) []byte { return msg.(*listener.TxListenerReceipt).ChainID.Bytes() })
+
 	// Pipe sarama message channel into worker
 	in := make(chan interface{})
 	go func() {
 		// Pipe channels for interface compatibility
-		pipeLoop:
+	pipeLoop:
 		for {
 			select {
 			case <-l.app.ctx.Done():
@@ -64,7 +66,7 @@ func (l *ListenerHandler) Setup() {
 
 	// Run worker
 	go l.worker.Run(in)
-	
+
 	// Start draining errors
 	go func() {
 		for err := range l.listener.Errors() {
@@ -73,7 +75,7 @@ func (l *ListenerHandler) Setup() {
 			}).WithError(err).Errorf("tx-listener: got error")
 		}
 	}()
-	
+
 	// Start draining blocks
 	go func() {
 		for block := range l.listener.Blocks() {
@@ -96,7 +98,7 @@ func (l *ListenerHandler) Listen() {
 		if err != nil {
 
 		}
-		l.listener.Listen(chainID, position.BlockNumber, position.TxIndex, l.cfg)
+		l.listener.Listen(chainID, position.BlockNumber, position.TxIndex)
 	}
 	// Wait for worker to be done
 	<-l.worker.Done()
@@ -129,15 +131,15 @@ func (l *ListenerHandler) getStartingPosition(chainID *big.Int) (*StartingPositi
 		// If we have never produced then we start from genesis
 		return &StartingPosition{}, nil
 	}
-	
+
 	// Parse last record using protobuffer
-	var pb tracepb.Trace
+	var pb trace.Trace
 	err = proto.Unmarshal(lastRecord.Value, &pb)
 	if err != nil {
 		return nil, err
 	}
-	
-	return &StartingPosition{int64(pb.Receipt.BlockNumber), int64(pb.Receipt.TxIndex)+1}, nil
+
+	return &StartingPosition{int64(pb.Receipt.BlockNumber), int64(pb.Receipt.TxIndex) + 1}, nil
 }
 
 // TranslateBlockNumber translate a starting block number into its integer value
@@ -225,76 +227,26 @@ func initListener(app *App) {
 	app.listener = &ListenerHandler{
 		app:             app,
 		startPositions:  positions,
-		defaultPosition: &StartingPosition{defaultPosition,0},
+		defaultPosition: &StartingPosition{defaultPosition, 0},
 		cleanOnce:       &sync.Once{},
-		cfg: config,
-		listener: listener.NewTxListener(listener.NewEthClient(app.infra.Mec, config)),
+		cfg:             config,
+		listener:        listener.NewTxListener(app.infra.Mec, config),
 	}
 	app.listener.Setup()
 }
 
 // InitFlags register flags for listener
 func InitFlags(f *pflag.FlagSet) {
-	ListenerBlockBackoff(f)
-	ListenerBlockLimit(f)
-	ListenerTrackerDepth(f)
+	listener.InitFlags(f)
 	ListenerStartDefault(f)
 	ListenerStart(f)
-}
-
-var (
-	listenerBlockBackoffFlag     = "listener-block-backoff"
-	listenerBlockBackoffViperKey = "listener.block.backoff"
-	listenerBlockBackoffDefault  = time.Second
-	listenerBlockBackoffEnv  = "LISTENER_BLOCK_BACKOFF"
-)
-
-// ListenerBlockBackoff register flag for Listener Block backoff
-func ListenerBlockBackoff(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Backoff time to wait before retrying after failing to find a mined block
-Environment variable: %q`, listenerBlockBackoffEnv)
-	f.Duration(listenerBlockBackoffFlag, listenerBlockBackoffDefault, desc)
-	viper.BindPFlag(listenerBlockBackoffViperKey, f.Lookup(listenerBlockBackoffFlag))
-	viper.BindEnv(listenerBlockBackoffViperKey, listenerBlockBackoffEnv)
-}
-
-var (
-	listenerBlockLimitFlag     = "listener-block-limit"
-	listenerBlockLimitViperKey = "listener.block.limit"
-	listenerBlockLimitDefault  = int64(40)
-	listenerBlockLimitEnv  = "LISTENER_BLOCK_LIMIT"
-)
-
-// ListenerBlockLimit register flag for Listener Block limit
-func ListenerBlockLimit(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Limit number of block that can be prefetched while listening
-Environment variable: %q`, listenerBlockLimitEnv)
-	f.Int64(listenerBlockLimitFlag, listenerBlockLimitDefault, desc)
-	viper.BindPFlag(listenerBlockLimitViperKey, f.Lookup(listenerBlockLimitFlag))
-	viper.BindEnv(listenerBlockLimitViperKey, listenerBlockLimitEnv)
-}
-
-var (
-	listenerTrackerDepthFlag     = "listener-tracker-depth"
-	listenerTrackerDepthViperKey = "listener.tracker.depth"
-	listenerTrackerDepthDefault  = int64(5)
-	listenerTrackerDepthEnv  = "LISTENER_TRACKER_DEPTH"
-)
-
-// ListenerTrackerDepth register flag for Listener Tracker Depth
-func ListenerTrackerDepth(f *pflag.FlagSet) {
-	desc := fmt.Sprintf(`Depth at which we consider a block final (to avoid falling into a re-org)
-Environment variable: %q`, listenerTrackerDepthEnv)
-	f.Int64(listenerTrackerDepthFlag, listenerTrackerDepthDefault, desc)
-	viper.BindPFlag(listenerTrackerDepthViperKey, f.Lookup(listenerTrackerDepthFlag))
-	viper.BindEnv(listenerTrackerDepthViperKey, listenerTrackerDepthEnv)
 }
 
 var (
 	listenerStartDefaultFlag     = "listener-start-default"
 	listenerStartDefaultViperKey = "listener.start.default"
 	listenerStartDefaultDefault  = "oldest"
-	listenerStartDefaultEnv  = "LISTENER_START_DEFAULT"
+	listenerStartDefaultEnv      = "LISTENER_START_DEFAULT"
 )
 
 // ListenerStartDefault register flag for Listener Start Default
@@ -310,7 +262,7 @@ var (
 	listenerStartFlag     = "listener-start"
 	listenerStartViperKey = "listener.start"
 	listenerStartDefault  = []string{}
-	listenerStartEnv  = "LISTENER_START"
+	listenerStartEnv      = "LISTENER_START"
 )
 
 // ListenerStart register flag for Listener Start Position