+++ cmd/run.go
@@ -5,8 +5,10 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
-	"gitlab.com/ConsenSys/client/fr/core-stack/common.git/config"
-	"gitlab.com/ConsenSys/client/fr/core-stack/common.git/utils"
+	storegrpc "gitlab.com/ConsenSys/client/fr/core-stack/api/context-store.git/infra/grpc"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/common/config"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/common/utils"
+	"gitlab.com/ConsenSys/client/fr/core-stack/pkg.git/core/worker"
 	"gitlab.com/ConsenSys/client/fr/core-stack/worker/tx-listener.git/app"
 )
 
@@ -23,6 +25,8 @@ func newRunCommand() *cobra.Command {
 	config.EthClientURLs(runCmd.Flags())
 	config.KafkaAddresses(runCmd.Flags())
 	config.TxDecoderOutTopic(runCmd.Flags())
+	worker.InitFlags(runCmd.Flags())
+	storegrpc.StoreTarget(runCmd.Flags())
 
 	return runCmd
 }